src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["3R11K <erikbatista1222@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.35.44"
numpy = "^2.1.2"
pandas = "^2.2.3"
jupyter = "^1.1.1"
pyarrow = "^17.0.0"
python-dotenv = "^1.0.1"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
openpyxl = "^3.1.5"
scikit-learn = "^1.5.2"
plotly = "^5.24.1"
pydantic = "^2.9.2"
clickhouse-driver = "^0.2.9"
psycopg2-binary = "^2.9.10"
flask = "^3.1.0"
clickhouse-connect = "^0.8.6"
flasgger = "^0.9.7.1"
prefect = "^3.1.4"
pytest-cov = "^6.0.0"
streamlit = "^1.40.2"
requests = "^2.32.3"
flask-cors = "^5.0.0"
streamlit-authenticator = "^0.4.1"
extra-streamlit-components = "^0.1.71"
aiohttp = "^3.11.10"
streamlit-cookies-controller = "^0.0.4"
gunicorn = "^23.0.0"



[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"